+++ lib/faker/deprecate/singular_siegler.rb
@@ -3,9 +3,13 @@
 module Faker
   class SingularSiegler < Base
     class << self
+      extend Gem::Deprecate
+
       def quote
-        fetch('singular_siegler.quotes')
+        Faker::Quote.singular_siegler
       end
+
+      deprecate :quote, 'Faker::Quote.singular_siegler', 2018, 12
     end
   end
 end