+++ lib/faker/deprecate/yoda.rb
@@ -3,10 +3,14 @@
 module Faker
   class Yoda < Base
     class << self
+      extend Gem::Deprecate
+
       # from: http://morecoolquotes.com/famous-yoda-quotes/
       def quote
-        fetch('yoda.quotes')
+        Faker::Quote.yoda
       end
+
+      deprecate :quote, 'Faker::Quote.yoda', 2018, 12
     end
   end
 end