+++ test/deprecate/test_faker_matz.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require_relative '../../test_helper'
+require_relative '../test_helper'
 
-class TestFakerMatz < Test::Unit::TestCase
+class TestDeprecateMatz < Test::Unit::TestCase
   def setup
     @tester = Faker::Matz
   end