+++ test/deprecate/test_most_interesting_man_in_the_world.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require_relative '../../test_helper'
+require_relative '../test_helper'
 
-class MostInterestingManInTheWorld < Test::Unit::TestCase
+class TestDeprecateMostInterestingManInTheWorld < Test::Unit::TestCase
   def setup
     @tester = Faker::MostInterestingManInTheWorld
   end